fix(github): add gateway sentinel mappings to ClassifyGitHubError (#32)
Extract ClassifyGitHubError from client.go into classify.go and add
errors.Is checks for all ErrGateway* sentinels introduced by #33.

Mapping:
- ErrGatewayRotationFailed  -> TOKEN_REFRESH_FAILED  (must re-auth)
- ErrGatewaySubjectGone     -> REAUTH_REQUIRED       (subject revoked)
- ErrGatewayUpstreamFailure -> TRANSIENT_UPSTREAM_ERROR (retryable)
- ErrGatewayUnauthorized    -> AUTH_REQUIRED          (config error)
- ErrGatewayLoopbackRequired-> AUTH_REQUIRED          (config error)
- ErrGatewayBadRequest      -> VALIDATION_ERROR       (config error)

Gateway sentinels are checked before generic HTTP-status checks because
they carry more precise semantics. Wrapped errors are also covered via
errors.Is unwrapping.
