+++ internal/github/client.go
@@ -14,8 +14,6 @@ import (
 	"github.com/google/go-github/v85/github"
 	"github.com/shurcooL/githubv4"
 	"golang.org/x/oauth2"
-
-	"github.com/scottlz0310/copilot-review-mcp/internal/autherr"
 )
 
 // invalidatingTransport wraps an HTTP transport and calls invalidate(token) when
@@ -342,78 +340,6 @@ func IsAuthError(err error) bool {
 	// shurcooL/githubv4 returns a plain error with the message
 	// "non-200 OK status code: 401 Unauthorized body: ..." for HTTP 401.
 	return strings.Contains(err.Error(), "401 Unauthorized")
-}
-
-// ClassifyGitHubError maps a GitHub API error to a structured *autherr.AuthError.
-// Returns nil when the error is not a recognized GitHub API error (e.g. context.Canceled).
-//
-// Detection order (most-specific first):
-//  1. *autherr.AuthError — already classified upstream; returned as-is.
-//  2. *github.RateLimitError — primary rate limit; retryable, safe to continue after reset.
-//  3. *github.AbuseRateLimitError — secondary / abuse rate limit; not retryable.
-//  4. *github.ErrorResponse — classified by HTTP status code.
-//  5. String-based fallback for shurcooL/githubv4 plain errors.
-func ClassifyGitHubError(err error) *autherr.AuthError {
-	if err == nil {
-		return nil
-	}
-
-	// Already classified.
-	if ae, ok := autherr.AsAuthError(err); ok {
-		return ae
-	}
-
-	// Primary rate limit (go-github wraps 403 + X-RateLimit-Remaining: 0).
-	var rateLimitErr *github.RateLimitError
-	if errors.As(err, &rateLimitErr) {
-		return autherr.NewRateLimited(true, true)
-	}
-
-	// Secondary / abuse rate limit (go-github wraps 403 + abuse message).
-	var abuseErr *github.AbuseRateLimitError
-	if errors.As(err, &abuseErr) {
-		return autherr.NewRateLimited(false, false)
-	}
-
-	// REST API errors with explicit HTTP status codes.
-	var ghErr *github.ErrorResponse
-	if errors.As(err, &ghErr) && ghErr.Response != nil {
-		switch ghErr.Response.StatusCode {
-		case http.StatusUnauthorized:
-			return autherr.NewReauthRequired()
-		case http.StatusForbidden:
-			return autherr.NewPermissionDenied()
-		case http.StatusNotFound:
-			return autherr.NewNotFound()
-		case http.StatusBadRequest, http.StatusUnprocessableEntity:
-			return autherr.NewValidationError()
-		case http.StatusInternalServerError,
-			http.StatusBadGateway,
-			http.StatusServiceUnavailable,
-			http.StatusGatewayTimeout:
-			return autherr.NewTransientUpstreamError()
-		}
-	}
-
-	// shurcooL/githubv4 plain-error fallbacks (string-based).
-	msg := err.Error()
-	switch {
-	case strings.Contains(msg, "401 Unauthorized"):
-		return autherr.NewReauthRequired()
-	case strings.Contains(msg, "403 Forbidden"):
-		return autherr.NewPermissionDenied()
-	case strings.Contains(msg, "404 Not Found"):
-		return autherr.NewNotFound()
-	case strings.Contains(msg, "422 Unprocessable"), strings.Contains(msg, "400 Bad Request"):
-		return autherr.NewValidationError()
-	case strings.Contains(msg, "500 Internal Server Error"),
-		strings.Contains(msg, "502 Bad Gateway"),
-		strings.Contains(msg, "503 Service Unavailable"),
-		strings.Contains(msg, "504 Gateway Timeout"):
-		return autherr.NewTransientUpstreamError()
-	}
-
-	return nil
 }
 
 // prNodeIDQuery fetches the GraphQL node ID for a pull request.
